+++ plugins/API/tests/JsonRendererTest.php
@@ -249,8 +249,9 @@ class JsonRendererTest extends \PHPUnit_Framework_TestCase
         $input = array('nb_visits' => 6, 'nb_random' => 8);
 
         $response = $this->jsonBuilder->renderArray($input);
+        $expected = json_encode($input);
 
-        $this->assertEquals('[{"nb_visits":6,"nb_random":8}]', $response);
+        $this->assertEquals($expected, $response);
         $this->assertNoJsonError($response);
     }
 
@@ -262,8 +263,9 @@ class JsonRendererTest extends \PHPUnit_Framework_TestCase
         );
 
         $response = $this->jsonBuilder->renderArray($input);
+        $expected = json_encode($input);
 
-        $this->assertEquals('[{"nb_visits":6,"nb_random":8},{"nb_visits":3,"nb_random":4}]', $response);
+        $this->assertEquals($expected, $response);
         $this->assertNoJsonError($response);
     }
 
@@ -292,6 +294,20 @@ class JsonRendererTest extends \PHPUnit_Framework_TestCase
                 "secondElement" => "isSecond",
             ),
             "thirdElement"  => "isThird");
+
+        $expected = json_encode($input);
+
+        $actual = $this->jsonBuilder->renderArray($input);
+        $this->assertEquals($expected, $actual);
+        $this->assertNoJsonError($actual);
+    }
+
+    public function test_renderArray_ShouldConvertSingleDimensionalAssociativeArrayToJson()
+    {
+        $input = array(
+            "fistElement" => "isFirst",
+            "secondElement" => "isSecond"
+        );
 
         $expected = json_encode($input);