+++ plugins/CoreHome/angularjs/menudropdown/menudropdown-directive.js
@@ -60,7 +60,7 @@ angular.module('piwikApp').directive('piwikMenudropdown', function(){
                         $node.show();
                     }
                 });
-            }
+            };
         }
     };
 });